# swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Clinic": {
				"properties": {
					"name": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"availability": {
						"properties": {
							"to": {
								"type": "string"
							},
							"from": {
								"type": "string"
							}
						},
						"required": [
							"to",
							"from"
						],
						"type": "object"
					}
				},
				"required": [
					"name",
					"state",
					"availability"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "node",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/clinic/search": {
			"get": {
				"operationId": "SearchClinics",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Clinic"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Internal Server Error",
											"statusCode": 500,
											"message": "Error Message"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}